mood_regulator/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
PyYAML==6.0.1
httpx==0.27.0
python-dotenv==1.0.1
aiofiles